CROPS-36 Filter recent posts by field_blog_author
We don't need to use the alias taxonomy here because
the relatinship is as follows:

blog post ---has one---> people ---has many ---> aliases

Therefore, we don't need to go deep to the alias relationship to
get content for a given people node.

This is a different case than the filter that we use for /blog, where
we want to filter by an alias and get the blog posts for the related
people node.